requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
boto3==1.34.34
pandas==2.2.0
psycopg2-binary==2.9.9
pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.27.0